Extract admin route registration
